Moves database into seperate folder
