+++ components/brave_rewards/browser/BUILD.gn
@@ -71,14 +71,14 @@ source_set("browser") {
 
   if (brave_rewards_enabled) {
     sources += [
+      "database/publisher_info_database.cc",
+      "database/publisher_info_database.h",
       "net/network_delegate_helper.cc",
       "net/network_delegate_helper.h",
       "rewards_service_impl.cc",
       "rewards_service_impl.h",
       "publisher_info_backend.cc",
       "publisher_info_backend.h",
-      "publisher_info_database.cc",
-      "publisher_info_database.h",
       "rewards_fetcher_service_observer.cc",
       "rewards_fetcher_service_observer.h",
       "rewards_notification_service_impl.cc",
