+++ components/brave_rewards/browser/database/publisher_info_database.h
