+++ test/BUILD.gn
@@ -142,7 +142,7 @@ test("brave_unit_tests") {
       "//brave/vendor/bat-native-ledger/src/bat/ledger/internal/bat_publishers_unittest.cc",
       "//brave/vendor/bat-native-ledger/src/bat/ledger/internal/bat_publishers_unittest.h",
       "//brave/vendor/bat-native-ledger/src/bat/ledger/internal/test/niceware_partial_unittest.cc",
-      "//brave/components/brave_rewards/browser/publisher_info_database_unittest.cc",
+      "//brave/components/brave_rewards/browser/database/publisher_info_database_unittest.cc",
       "//brave/components/brave_rewards/browser/rewards_service_impl_unittest.cc",
       "//brave/vendor/bat-native-ads/src/bat/ads/internal/ads_is_mobile_unittest.cc",
       "//brave/vendor/bat-native-ads/src/bat/ads/internal/ads_tabs_unittest.cc",
